+++ processor/src/test/fixtures/expected/com/example/arez/NonTrackingHasActionBuilder.java
@@ -1,0 +1,47 @@
+package com.example.arez;
+
+import javax.annotation.Nonnull;
+import javax.annotation.processing.Generated;
+import org.jetbrains.annotations.Contract;
+import react4j.ReactElement;
+import react4j.ReactNode;
+
+@Generated("react4j.processor.React4jProcessor")
+final class NonTrackingHasActionBuilder {
+  private NonTrackingHasActionBuilder() {
+  }
+
+  @Nonnull
+  static Step1 newBuilder() {
+    return new Builder();
+  }
+
+  @Nonnull
+  @Contract(
+      pure = true
+  )
+  static ReactNode build() {
+    return newBuilder().build();
+  }
+
+  interface Step1 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    ReactNode build();
+  }
+
+  private static class Builder implements Step1 {
+    @Nonnull
+    private final ReactElement _element = ReactElement.createViewElement( React4j_NonTrackingHasActionView.Factory.TYPE );
+
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final ReactNode build() {
+      return _element;
+    }
+  }
+}