+++ processor/src/test/fixtures/expected/com/example/default_inputs/PackageAccessFieldInputDefaultBuilder.java
@@ -1,0 +1,79 @@
+package com.example.default_inputs;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import javax.annotation.processing.Generated;
+import jsinterop.base.JsPropertyMap;
+import org.jetbrains.annotations.Contract;
+import react4j.ReactElement;
+import react4j.ReactNode;
+
+@Generated("react4j.processor.React4jProcessor")
+final class PackageAccessFieldInputDefaultBuilder {
+  private PackageAccessFieldInputDefaultBuilder() {
+  }
+
+  @Nonnull
+  static Step1 newBuilder() {
+    return new Builder();
+  }
+
+  @Nonnull
+  @Contract(
+      pure = true
+  )
+  static ReactNode myProp(@Nullable final String myProp) {
+    return newBuilder().myProp( myProp );
+  }
+
+  @Nonnull
+  @Contract(
+      pure = true
+  )
+  static ReactNode build() {
+    return newBuilder().build();
+  }
+
+  interface Step1 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    ReactNode myProp(@Nullable String myProp);
+
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    ReactNode build();
+  }
+
+  private static class Builder implements Step1 {
+    @Nonnull
+    private final ReactElement _element;
+
+    Builder() {
+      _element = ReactElement.createViewElement( React4j_PackageAccessFieldInputDefaultView.Factory.TYPE );
+      final JsPropertyMap<Object> inputs = _element.inputs();
+      inputs.set( React4j_PackageAccessFieldInputDefaultView.Inputs.myProp, PackageAccessFieldInputDefaultView.DEFAULT_MY_PROP );
+    }
+
+    @Override
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final ReactNode myProp(@Nullable final String myProp) {
+      _element.input( React4j_PackageAccessFieldInputDefaultView.Inputs.myProp, myProp );
+      return build();
+    }
+
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final ReactNode build() {
+      return _element;
+    }
+  }
+}