+++ processor/src/test/fixtures/expected/com/example/inheritance/simple/LeafBuilder.java
@@ -1,0 +1,266 @@
+package com.example.inheritance.simple;
+
+import javax.annotation.Nonnull;
+import javax.annotation.processing.Generated;
+import org.jetbrains.annotations.Contract;
+import react4j.ReactElement;
+import react4j.ReactNode;
+
+@Generated("react4j.processor.React4jProcessor")
+final class LeafBuilder {
+  private LeafBuilder() {
+  }
+
+  @Nonnull
+  static Step1 newBuilder() {
+    return new Builder();
+  }
+
+  @Nonnull
+  @Contract(
+      pure = true
+  )
+  static Step2 zzField1InRootInterface(final boolean zzField1InRootInterface) {
+    return newBuilder().zzField1InRootInterface( zzField1InRootInterface );
+  }
+
+  interface Step1 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    Step2 zzField1InRootInterface(boolean zzField1InRootInterface);
+  }
+
+  interface Step2 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    Step3 aaField2InRootInterface(boolean aaField2InRootInterface);
+  }
+
+  interface Step3 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    Step4 field1InRoot(boolean field1InRoot);
+  }
+
+  interface Step4 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    Step5 field2InRoot(boolean field2InRoot);
+  }
+
+  interface Step5 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    Step6 zzField1InInnerInterface(boolean zzField1InInnerInterface);
+  }
+
+  interface Step6 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    Step7 aaField2InInnerInterface(boolean aaField2InInnerInterface);
+  }
+
+  interface Step7 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    Step8 field1InInner(boolean field1InInner);
+  }
+
+  interface Step8 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    Step9 field2InInner(boolean field2InInner);
+  }
+
+  interface Step9 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    Step10 zzField1InLeafInterface(boolean zzField1InLeafInterface);
+  }
+
+  interface Step10 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    Step11 aaField2InLeafInterface(boolean aaField2InLeafInterface);
+  }
+
+  interface Step11 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    Step12 field1InLeaf(boolean field1InLeaf);
+  }
+
+  interface Step12 {
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    ReactNode field2InLeaf(boolean field2InLeaf);
+  }
+
+  private static class Builder implements Step1,
+      Step2,
+      Step3,
+      Step4,
+      Step5,
+      Step6,
+      Step7,
+      Step8,
+      Step9,
+      Step10,
+      Step11,
+      Step12 {
+    @Nonnull
+    private final ReactElement _element = ReactElement.createViewElement( React4j_LeafView.Factory.TYPE );
+
+    @Override
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final Step2 zzField1InRootInterface(final boolean zzField1InRootInterface) {
+      _element.input( React4j_LeafView.Inputs.zzField1InRootInterface, zzField1InRootInterface );
+      return this;
+    }
+
+    @Override
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final Step3 aaField2InRootInterface(final boolean aaField2InRootInterface) {
+      _element.input( React4j_LeafView.Inputs.aaField2InRootInterface, aaField2InRootInterface );
+      return this;
+    }
+
+    @Override
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final Step4 field1InRoot(final boolean field1InRoot) {
+      _element.input( React4j_LeafView.Inputs.field1InRoot, field1InRoot );
+      return this;
+    }
+
+    @Override
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final Step5 field2InRoot(final boolean field2InRoot) {
+      _element.input( React4j_LeafView.Inputs.field2InRoot, field2InRoot );
+      return this;
+    }
+
+    @Override
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final Step6 zzField1InInnerInterface(final boolean zzField1InInnerInterface) {
+      _element.input( React4j_LeafView.Inputs.zzField1InInnerInterface, zzField1InInnerInterface );
+      return this;
+    }
+
+    @Override
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final Step7 aaField2InInnerInterface(final boolean aaField2InInnerInterface) {
+      _element.input( React4j_LeafView.Inputs.aaField2InInnerInterface, aaField2InInnerInterface );
+      return this;
+    }
+
+    @Override
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final Step8 field1InInner(final boolean field1InInner) {
+      _element.input( React4j_LeafView.Inputs.field1InInner, field1InInner );
+      return this;
+    }
+
+    @Override
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final Step9 field2InInner(final boolean field2InInner) {
+      _element.input( React4j_LeafView.Inputs.field2InInner, field2InInner );
+      return this;
+    }
+
+    @Override
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final Step10 zzField1InLeafInterface(final boolean zzField1InLeafInterface) {
+      _element.input( React4j_LeafView.Inputs.zzField1InLeafInterface, zzField1InLeafInterface );
+      return this;
+    }
+
+    @Override
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final Step11 aaField2InLeafInterface(final boolean aaField2InLeafInterface) {
+      _element.input( React4j_LeafView.Inputs.aaField2InLeafInterface, aaField2InLeafInterface );
+      return this;
+    }
+
+    @Override
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final Step12 field1InLeaf(final boolean field1InLeaf) {
+      _element.input( React4j_LeafView.Inputs.field1InLeaf, field1InLeaf );
+      return this;
+    }
+
+    @Override
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final ReactNode field2InLeaf(final boolean field2InLeaf) {
+      _element.input( React4j_LeafView.Inputs.field2InLeaf, field2InLeaf );
+      return build();
+    }
+
+    @Nonnull
+    @Contract(
+        pure = true
+    )
+    public final ReactNode build() {
+      return _element;
+    }
+  }
+}